requirements.txt
anthropic
numpy
opencv-python
Pillow
ultralytics
sentence-transformers
streamlit
pinecone-client
